Make sure theta axis type float
And some misc code cleanup along the way.

Eliminate the original tiling_wizard_controller, since we have not been using this for a while.

Rename second tiling_wizard_controller (e.g., remove the 2 at the end).

Some minor changes in formatting.
